+++ Sources/Sentry/SentryCrashIntegration.m
@@ -140,6 +140,7 @@ SentryCrashIntegration ()
             [SentryCrashIntegration sendAllSentryCrashReports];
         }
     };
+    [self.crashAdapter startBinaryImageCache];
     [self.dispatchQueueWrapper dispatchOnce:&installationToken block:block];
 }
 
@@ -157,6 +158,8 @@ SentryCrashIntegration ()
         [installation uninstall];
         installationToken = 0;
     }
+
+    [self.crashAdapter stopBinaryImageCache];
 
     [NSNotificationCenter.defaultCenter removeObserver:self
                                                   name:NSCurrentLocaleDidChangeNotification